This leaves range.rs alone.

+++ components/util/opts.rs
@@ -28,14 +28,14 @@ pub struct Opts {
     /// How many threads to use for CPU painting (`-t`).
     ///
     /// Note that painting is sequentialized when using GPU painting.
-    pub paint_threads: uint,
+    pub paint_threads: usize,
 
     /// True to use GPU painting via Skia-GL, false to use CPU painting via Skia (`-g`). Note that
     /// compositing is always done on the GPU.
     pub gpu_painting: bool,
 
     /// The maximum size of each tile in pixels (`-s`).
-    pub tile_size: uint,
+    pub tile_size: usize,
 
     /// The ratio of device pixels per px at the default scale. If unspecified, will use the
     /// platform default setting.
@@ -54,7 +54,7 @@ pub struct Opts {
 
     /// The number of threads to use for layout (`-y`). Defaults to 1, which results in a recursive
     /// sequential algorithm.
-    pub layout_threads: uint,
+    pub layout_threads: usize,
 
     pub nonincremental_layout: bool,
 
@@ -102,7 +102,7 @@ pub struct Opts {
     pub devtools_port: Option<u16>,
 
     /// The initial requested size of the window.
-    pub initial_window_size: TypedSize2D<ScreenPx, uint>,
+    pub initial_window_size: TypedSize2D<ScreenPx, u32>,
 
     /// An optional string allowing the user agent to be set for testing.
     pub user_agent: Option<String>,
@@ -256,7 +256,7 @@ pub fn from_cmdline_args(args: &[String]) -> bool {
         opt_match.free.clone()
     };
 
-    let tile_size: uint = match opt_match.opt_str("s") {
+    let tile_size: usize = match opt_match.opt_str("s") {
         Some(tile_size_str) => tile_size_str.parse().unwrap(),
         None => 512,
     };
@@ -265,7 +265,7 @@ pub fn from_cmdline_args(args: &[String]) -> bool {
         ScaleFactor(dppx_str.parse().unwrap())
     );
 
-    let mut paint_threads: uint = match opt_match.opt_str("t") {
+    let mut paint_threads: usize = match opt_match.opt_str("t") {
         Some(paint_threads_str) => paint_threads_str.parse().unwrap(),
         None => cmp::max(rt::default_sched_threads() * 3 / 4, 1),
     };
@@ -280,7 +280,7 @@ pub fn from_cmdline_args(args: &[String]) -> bool {
 
     let gpu_painting = !FORCE_CPU_PAINTING && opt_match.opt_present("g");
 
-    let mut layout_threads: uint = match opt_match.opt_str("y") {
+    let mut layout_threads: usize = match opt_match.opt_str("y") {
         Some(layout_threads_str) => layout_threads_str.parse().unwrap(),
         None => cmp::max(rt::default_sched_threads() * 3 / 4, 1),
     };
@@ -301,7 +301,7 @@ pub fn from_cmdline_args(args: &[String]) -> bool {
 
     let initial_window_size = match opt_match.opt_str("resolution") {
         Some(res_string) => {
-            let res: Vec<uint> = res_string.split('x').map(|r| r.parse().unwrap()).collect();
+            let res: Vec<u32> = res_string.split('x').map(|r| r.parse().unwrap()).collect();
             TypedSize2D(res[0], res[1])
         }
         None => {
